refactor: split operator user management services
